args: Allow the use of callbacks after argument compilation
In some cases, it is necessary to perform some actions after invoking
gccrs. For example, when generating a static library, you need to invoke
the `ar` command and create an archive containing the object
files/compilation units you just generated. This also paves the way for
parsing the output of `gccrs -frust-dump-target_options` once this part
will be handled by the options parser as well and not separately