object_file_name: Set extension instead of adding an extra extension to object file name
The previous implementation would try to add a new component to the
cloned PathBuf, which would act as if the original output file was a
directory. This is not the correct behavior. Instead, set a different
extension to the cloned PathBuf so that `lib<name>.a` becomes
`lib<name>.tmp.o`, which is basically the same